# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AyurvedicZen API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Map of question ID to selected answer value (a, b, or c)
 */
export type SubmitAnswersBodyAnswers = { [key: string]: string };

export interface SubmitAnswersBody {
  /** Map of question ID to selected answer value (a, b, or c) */
  answers: SubmitAnswersBodyAnswers;
}

export type DiagnosisResultResult =
  (typeof DiagnosisResultResult)[keyof typeof DiagnosisResultResult];

export const DiagnosisResultResult = {
  Vata: "Vata",
  Pitta: "Pitta",
  Kapha: "Kapha",
} as const;

export interface DiagnosisResult {
  id: number;
  result: DiagnosisResultResult;
  vataScore: number;
  pittaScore: number;
  kaphaScore: number;
  description: string;
  recommendations: string[];
  createdAt: string;
}

export type StoredResponseAnswers = { [key: string]: string };

export interface StoredResponse {
  id: number;
  answers: StoredResponseAnswers;
  result: string;
  vataScore: number;
  pittaScore: number;
  kaphaScore: number;
  createdAt: string;
}

export interface DiagnosisStats {
  total: number;
  vataCount: number;
  pittaCount: number;
  kaphaCount: number;
}

export interface ErrorResponse {
  error: string;
}