# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AyurvedicZen API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Receives user answers, analyzes them, and returns the Ayurvedic diagnosis
 * @summary Submit questionnaire answers
 */
export const SubmitAnswersBody = zod.object({
  answers: zod
    .record(zod.string(), zod.string())
    .describe("Map of question ID to selected answer value (a, b, or c)"),
});

/**
 * Returns all stored user responses and their diagnoses
 * @summary Get all past responses
 */
export const GetResponsesResponseItem = zod.object({
  id: zod.number(),
  answers: zod.record(zod.string(), zod.string()),
  result: zod.string(),
  vataScore: zod.number(),
  pittaScore: zod.number(),
  kaphaScore: zod.number(),
  createdAt: zod.string(),
});
export const GetResponsesResponse = zod.array(GetResponsesResponseItem);

/**
 * Returns a single response by ID
 * @summary Get a specific response
 */
export const GetResponseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResponseResponse = zod.object({
  id: zod.number(),
  answers: zod.record(zod.string(), zod.string()),
  result: zod.string(),
  vataScore: zod.number(),
  pittaScore: zod.number(),
  kaphaScore: zod.number(),
  createdAt: zod.string(),
});

/**
 * Returns counts of each dosha type diagnosed
 * @summary Get diagnosis statistics
 */
export const GetDiagnosisStatsResponse = zod.object({
  total: zod.number(),
  vataCount: zod.number(),
  pittaCount: zod.number(),
  kaphaCount: zod.number(),
});