# lib/api-zod/src/generated/types/diagnosisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AyurvedicZen API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisResultResult } from "./diagnosisResultResult";

export interface DiagnosisResult {
  id: number;
  result: DiagnosisResultResult;
  vataScore: number;
  pittaScore: number;
  kaphaScore: number;
  description: string;
  recommendations: string[];
  createdAt: string;
}