# lib/api-zod/src/generated/types/diagnosisStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AyurvedicZen API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnosisStats {
  total: number;
  vataCount: number;
  pittaCount: number;
  kaphaCount: number;
}