# lib/api-zod/src/generated/types/storedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AyurvedicZen API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoredResponseAnswers } from "./storedResponseAnswers";

export interface StoredResponse {
  id: number;
  answers: StoredResponseAnswers;
  result: string;
  vataScore: number;
  pittaScore: number;
  kaphaScore: number;
  createdAt: string;
}