# lib/api-zod/src/generated/types/submitAnswersBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AyurvedicZen API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitAnswersBodyAnswers } from "./submitAnswersBodyAnswers";

export interface SubmitAnswersBody {
  /** Map of question ID to selected answer value (a, b, or c) */
  answers: SubmitAnswersBodyAnswers;
}